requirements_raptor_enhanced.txt
# RAPTOR Enhanced - Dependências Organizadas
# Versão otimizada com embeddings reais e clustering avançado

# ===================================================================
# CORE DEPENDENCIES (Obrigatórias)
# ===================================================================

# Computação numérica e ML básico
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=1.5.0

# Async e HTTP
asyncio-throttle>=1.0.2
aiohttp>=3.9.0
httpx>=0.26.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
rich>=13.0.0
tqdm>=4.65.0

# ===================================================================
# EMBEDDING PROVIDERS (Pelo menos um recomendado)
# ===================================================================

# OpenAI (Recomendado - melhor qualidade)
openai>=1.0.0
tiktoken>=0.5.0

# Sentence Transformers (Alternativa local)
sentence-transformers>=2.2.0
torch>=2.0.0

# HuggingFace
transformers>=4.30.0

# ===================================================================
# CLUSTERING AVANÇADO (Opcional mas recomendado)
# ===================================================================

# UMAP para dimensionality reduction
umap-learn>=0.5.4

# Clustering adicional
hdbscan>=0.8.29

# ===================================================================
# LLM PROVIDERS (Para summarização)
# ===================================================================

# Anthropic Claude
anthropic>=0.25.0

# Google Gemini
google-generativeai>=0.3.0

# ===================================================================
# VECTOR DATABASES (Opcional)
# ===================================================================

# Qdrant (Recomendado)
qdrant-client>=1.8.0

# Chroma (Alternativa leve)
chromadb>=0.4.0

# Pinecone (Cloud)
pinecone-client>=2.2.0

# ===================================================================
# CACHE E PERFORMANCE (Opcional)
# ===================================================================

# Redis para cache distribuído
redis>=5.0.0

# Cache local
cachetools>=5.3.0

# Processamento paralelo
joblib>=1.3.0

# ===================================================================
# MONITORING E LOGGING (Opcional)
# ===================================================================

# Structured logging
structlog>=23.0.0
python-json-logger>=2.0.0

# Metrics
prometheus-client>=0.19.0

# ===================================================================
# DEVELOPMENT E TESTING (Dev only)
# ===================================================================

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
pytest-cov>=4.1.0

# Code quality
black>=23.0.0
ruff>=0.1.0
mypy>=1.5.0

# ===================================================================
# EXTRAS OPCIONAIS
# ===================================================================

# Text processing
spacy>=3.6.0
nltk>=3.8.0

# Document parsing
pypdf>=3.15.0
python-docx>=0.8.11
beautifulsoup4>=4.12.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Jupyter (para notebooks)
jupyter>=1.0.0
ipywidgets>=8.0.0

# ===================================================================
# INSTALAÇÃO RECOMENDADA
# ===================================================================

# Para instalação completa:
# pip install -r requirements_raptor_enhanced.txt

# Para instalação mínima:
# pip install numpy scikit-learn openai tiktoken sentence-transformers umap-learn

# Para desenvolvimento:
# pip install -r requirements_raptor_enhanced.txt
# pip install -e .

# ===================================================================
# NOTAS DE COMPATIBILIDADE
# ===================================================================

# Python 3.8+ requerido
# CUDA opcional para aceleração GPU (torch, sentence-transformers)
# Redis server opcional para cache distribuído
# Dependências de sistema podem ser necessárias para umap-learn 